Renamed to loadgenerator and moved under tools
